services/inventory-service/requirements.txt
Django>=5.2
djangorestframework
psycopg2-binary
python-dotenv
dj-database-url
